requirements.txt
# Home Assistant Core
aiohttp==3.8.1
astral==2.2
async_timeout==4.0.2
attrs==21.2.0
atomicwrites-homeassistant==1.4.1
awesomeversion==22.9.0
bcrypt==3.1.7
certifi>=2021.5.30
ciso8601==2.2.0
httpx==0.23.1
home-assistant-bluetooth==1.8.1
ifaddr==0.1.7
jinja2==3.1.2
lru-dict==1.1.8
PyJWT==2.5.0
cryptography==38.0.3
orjson==3.8.1
pip>=21.0,<22.4
python-slugify==4.0.1
pyyaml==6.0
requests==2.28.1
typing-extensions>=4.4.0,<5.0
voluptuous==0.13.1
voluptuous-serialize==2.5.0
yarl==1.8.1
pre-commit
flake8
isort
black
aioasuswrt==1.4.0
pyasuswrt==0.1.13
